MS_TelegramBot/requirements.txt
aiogram==3.4.1
cherrypy==18.8.0
requests==2.31.0
python-dotenv==1.0.1
paho-mqtt==2.2.1
aiohttp==3.9.3
typing-extensions==4.10.0